Making buffer be cleared by function and renaming functions folder
