+++ custom_functions.h
@@ -1,12 +1,16 @@
+#include "./constants.h"
+#include "./global_variables.h"
 #include "./structures.h"
 
 char* allocateMemoryForRequestMessage(char* http_method, char *data_to_send, char* hostname, char *html_file_path_and_filename_on_host);
 int calculateNewBufferSize();
+void cleanBufferOfDestinationFile(char (*buffer)[BUFFSIZ]);
 void configureSocket(struct addrinfo *address_info_configuration_model);
 char* copyDynamicString(char* output_str, char* input_str);
 char* createRequestMessage(char* http_method, char *data_to_send, char* hostname, char *html_file_path_and_filename_on_host);
 void eraseAllPreviousSocketData(struct addrinfo *socket_addr);
 void estabilishDataToSend(int socket_identificator, struct VariablesDTO variables);
+void forkAndLoopListThoTryToConnect(struct addrinfo *list_of_addresses_infos, int *socket_identificator);
 char* formatAsIP(struct sockaddr_in *information_core);
 struct VariablesDTO mapArgumentsToVariables(char *arguments[]);
 void setIfFamilyAddressIsIpv4OrIpv6(struct addrinfo *socket_addr, char protocol[]);
@@ -20,6 +24,10 @@ char* allocateMemoryForRequestMessage(char* http_method, char *data_to_send, cha
 
 int calculateNewBufferSize(){
     return BUFFSIZ - 1;
+}
+
+void cleanBufferOfDestinationFile(char (*buffer)[BUFFSIZ]){
+    bzero(buffer, BUFFSIZ);
 }
 
 void configureSocket(struct addrinfo *address_info_configuration_model){
@@ -69,6 +77,40 @@ void estabilishDataToSend(int socket_identificator, struct VariablesDTO variable
     
     //write on socket communication buffer
     write(socket_identificator, data_to_send, strlen(data_to_send));
+}
+
+void forkAndLoopListThoTryToConnect(struct addrinfo *list_of_addresses_infos, int *socket_identificator){
+    struct addrinfo *pointer_to_walk_along_address_infos;
+    struct sockaddr_in *information_core;
+    int OPTLEN = 1, UNCONNECTED_SOCKET = -1;
+
+    for(pointer_to_walk_along_address_infos = list_of_addresses_infos; pointer_to_walk_along_address_infos != NULL; pointer_to_walk_along_address_infos = pointer_to_walk_along_address_infos->ai_next)
+    {
+        information_core = (struct sockaddr_in *) pointer_to_walk_along_address_infos->ai_addr;
+        printf("------------------------\n");
+        printf("Address: %s\n", formatAsIP(information_core));
+        
+        int socket_identificator = socket(pointer_to_walk_along_address_infos->ai_family, pointer_to_walk_along_address_infos->ai_socktype, IPPROTO_TCP);
+
+        setsockopt(socket_identificator, IPPROTO_TCP, TCP_NODELAY, (const char *) &OPTLEN, sizeof(int));
+        
+        global_variable_to_get_socket_identificator_from_blocking_function_fork_and_loop_list_tho_try_to_connect = &socket_identificator;
+
+        if(socket_identificator == UNCONNECTED_SOCKET)
+            continue;
+        if(connect(socket_identificator, pointer_to_walk_along_address_infos->ai_addr, pointer_to_walk_along_address_infos->ai_addrlen) != -1)
+            break;
+        close(socket_identificator);
+
+    }
+
+    if (pointer_to_walk_along_address_infos == NULL)
+    {
+        fprintf(stderr, "It was not possible to estabilish connection\n");
+        exit(EXIT_FAILURE);
+    }
+
+    freeaddrinfo(list_of_addresses_infos);
 }
 
 char* formatAsIP(struct sockaddr_in *information_core){
